New auto config
Users can now toggle the main plugin functionality on and off. When this config is off, no auto-locking will occur. When it is on, the plugin will attempt to auto-lock align whenever char state changes to active and when a cskill cast attempt fails, assuming that the cskill is not already active.